--- a/pkg/rejectif/caa.go
+++ b/pkg/rejectif/caa.go
@@ -9,14 +9,6 @@ import (
 
 // Keep these in alphabetical order.
 
-// CaaTargetHasSemicolon identifies CAA records that contain semicolons.
-func CaaTargetHasSemicolon(rc *models.RecordConfig) error {
-	if strings.Contains(rc.GetTargetField(), ";") {
-		return fmt.Errorf("caa target contains semicolon")
-	}
-	return nil
-}
-
 // CaaFlagIsNonZero identifies CAA records where tag is no zero.
 func CaaFlagIsNonZero(rc *models.RecordConfig) error {
 	if rc.CaaFlag != 0 {
@@ -31,6 +23,14 @@ func CaaFlagIsNonZero(rc *models.RecordConfig) error {
 func CaaTargetContainsWhitespace(rc *models.RecordConfig) error {
 	if strings.ContainsAny(rc.GetTargetField(), " \t\r\n") {
 		return fmt.Errorf("caa target contains whitespace")
+	}
+	return nil
+}
+
+// CaaTargetHasSemicolon identifies CAA records that contain semicolons.
+func CaaTargetHasSemicolon(rc *models.RecordConfig) error {
+	if strings.Contains(rc.GetTargetField(), ";") {
+		return fmt.Errorf("caa target contains semicolon")
 	}
 	return nil
 }