--- a/pkg/rejectif/mx.go
+++ b/pkg/rejectif/mx.go
@@ -1,0 +1,18 @@
+package rejectif
+
+import (
+	"fmt"
+
+	"github.com/StackExchange/dnscontrol/v3/models"
+)
+
+// Keep these in alphabetical order.
+
+// MxNull detects MX records that are a "null MX".
+// This is needed by providers that don't support RFC 7505.
+func MxNull(rc *models.RecordConfig) error {
+	if rc.GetTargetField() == "." {
+		return fmt.Errorf("mx has null target")
+	}
+	return nil
+}