--- a/providers/cscglobal/auditrecords.go
+++ b/providers/cscglobal/auditrecords.go
@@ -19,6 +19,8 @@ func AuditRecords(records []*models.RecordConfig) []error {
 
 	a.Add("TXT", rejectif.TxtHasDoubleQuotes) // Last verified 2022-08-08
 
+	a.Add("SRV", rejectif.SrvHasNullTarget) // Last verified 2020-12-28
+
 	return a.Audit(records)
 }