--- a/providers/digitalocean/auditrecords.go
+++ b/providers/digitalocean/auditrecords.go
@@ -13,9 +13,11 @@ import (
 func AuditRecords(records []*models.RecordConfig) []error {
 	a := rejectif.Auditor{}
 
-	a.Add("TXT", MaxLengthDO) // Last verified 2021-03-01
-
 	a.Add("CAA", rejectif.CaaTargetHasSemicolon) // Last verified 2021-03-01
+
+	a.Add("CAA", rejectif.CaaTargetContainsWhitespace) // Last verified xxxx-xx-xx
+
+	a.Add("TXT", MaxLengthDO) // Last verified 2021-03-01
 
 	a.Add("TXT", rejectif.TxtHasDoubleQuotes) // Last verified 2021-03-01
 	// Double-quotes not permitted in TXT strings. I have a hunch that