--- a/providers/digitalocean/auditrecords.go
+++ b/providers/digitalocean/auditrecords.go
@@ -13,9 +13,11 @@ import (
 func AuditRecords(records []*models.RecordConfig) []error {
 	a := rejectif.Auditor{}
 
+	a.Add("CAA", rejectif.CaaTargetContainsWhitespace) // Last verified xxxx-xx-xx
+
 	a.Add("CAA", rejectif.CaaTargetHasSemicolon) // Last verified 2021-03-01
 
-	a.Add("CAA", rejectif.CaaTargetContainsWhitespace) // Last verified xxxx-xx-xx
+	a.Add("MX", rejectif.MxNull) // Last verified 2020-12-28
 
 	a.Add("TXT", MaxLengthDO) // Last verified 2021-03-01