--- a/providers/dnsimple/auditrecords.go
+++ b/providers/dnsimple/auditrecords.go
@@ -13,14 +13,16 @@ func AuditRecords(records []*models.RecordConfig) []error {
 
 	a.Add("CAA", rejectif.CaaTargetContainsWhitespace) // Last verified xxxx-xx-xx
 
+	a.Add("MX", rejectif.MxNull) // Last verified 2020-12-28
+
 	a.Add("TXT", rejectif.TxtHasMultipleSegments) // Last verified 2022-07
 	//TODO(onlyhavecans) I think we can support multiple strings.
 
 	a.Add("TXT", rejectif.TxtHasTrailingSpace) // Last verified 2022-07
 
-	a.Add("TXT", rejectif.TxtIsEmpty) // Last verified 2022-07
-
 	a.Add("TXT", rejectif.TxtHasUnpairedDoubleQuotes) // Last verified 2022-07
+
+	a.Add("TXT", rejectif.TxtIsEmpty) // Last verified 2022-07
 
 	return a.Audit(records)
 }