--- a/providers/exoscale/auditrecords.go
+++ b/providers/exoscale/auditrecords.go
@@ -1,10 +1,19 @@
 package exoscale
 
-import "github.com/StackExchange/dnscontrol/v3/models"
+import (
+	"github.com/StackExchange/dnscontrol/v3/models"
+	"github.com/StackExchange/dnscontrol/v3/pkg/rejectif"
+)
 
 // AuditRecords returns a list of errors corresponding to the records
 // that aren't supported by this provider.  If all records are
 // supported, an empty list is returned.
 func AuditRecords(records []*models.RecordConfig) []error {
-	return nil
+	a := rejectif.Auditor{}
+
+	a.Add("CAA", rejectif.CaaTargetContainsWhitespace) // Last verified 2022-07-11
+
+	a.Add("SRV", rejectif.SrvHasNullTarget) // Last verified 2020-12-28
+
+	return a.Audit(records)
 }