--- a/providers/exoscale/auditrecords.go
+++ b/providers/exoscale/auditrecords.go
@@ -13,6 +13,8 @@ func AuditRecords(records []*models.RecordConfig) []error {
 
 	a.Add("CAA", rejectif.CaaTargetContainsWhitespace) // Last verified 2022-07-11
 
+	a.Add("MX", rejectif.MxNull) // Last verified 2022-07-11
+
 	a.Add("SRV", rejectif.SrvHasNullTarget) // Last verified 2020-12-28
 
 	return a.Audit(records)