--- a/providers/hexonet/auditrecords.go
+++ b/providers/hexonet/auditrecords.go
@@ -13,5 +13,7 @@ func AuditRecords(records []*models.RecordConfig) []error {
 
 	a.Add("TXT", rejectif.TxtIsEmpty) // Last verified 2021-10-01
 
+	a.Add("SRV", rejectif.SrvHasNullTarget) // Last verified 2020-12-28
+
 	return a.Audit(records)
 }