--- a/providers/linode/auditrecords.go
+++ b/providers/linode/auditrecords.go
@@ -1,10 +1,18 @@
 package linode
 
-import "github.com/StackExchange/dnscontrol/v3/models"
+import (
+	"github.com/StackExchange/dnscontrol/v3/models"
+	"github.com/StackExchange/dnscontrol/v3/pkg/rejectif"
+)
 
 // AuditRecords returns a list of errors corresponding to the records
 // that aren't supported by this provider.  If all records are
 // supported, an empty list is returned.
 func AuditRecords(records []*models.RecordConfig) []error {
-	return nil
+
+	a := rejectif.Auditor{}
+
+	a.Add("CAA", rejectif.CaaFlagIsNonZero) // Last verified 2022-03-25
+
+	return a.Audit(records)
 }