--- a/providers/msdns/auditrecords.go
+++ b/providers/msdns/auditrecords.go
@@ -23,5 +23,7 @@ func AuditRecords(records []*models.RecordConfig) []error {
 
 	a.Add("TXT", rejectif.TxtIsEmpty) // Last verified 2021-03-01
 
+	a.Add("SRV", rejectif.SrvHasNullTarget) // Last verified 2020-12-28
+
 	return a.Audit(records)
 }