--- a/providers/msdns/auditrecords.go
+++ b/providers/msdns/auditrecords.go
@@ -11,19 +11,21 @@ import (
 func AuditRecords(records []*models.RecordConfig) []error {
 	a := rejectif.Auditor{}
 
+	a.Add("MX", rejectif.MxNull) // Last verified 2020-12-28
+
+	a.Add("SRV", rejectif.SrvHasNullTarget) // Last verified 2020-12-28
+
 	a.Add("TXT", rejectif.TxtHasBackticks) // Last verified 2021-03-01
 
 	a.Add("TXT", rejectif.TxtHasDoubleQuotes) // Last verified 2021-03-01
 
 	a.Add("TXT", rejectif.TxtHasMultipleSegments) // Last verified 2021-03-01
 
-	a.Add("TXT", rejectif.TxtHasSingleQuotes) // Last verified 2021-03-01
-
 	a.Add("TXT", rejectif.TxtHasSegmentLen256orLonger)
 
-	a.Add("TXT", rejectif.TxtIsEmpty) // Last verified 2021-03-01
+	a.Add("TXT", rejectif.TxtHasSingleQuotes) // Last verified 2021-03-01
 
-	a.Add("SRV", rejectif.SrvHasNullTarget) // Last verified 2020-12-28
+	a.Add("TXT", rejectif.TxtIsEmpty) // Last verified 2021-03-01
 
 	return a.Audit(records)
 }