--- a/providers/namedotcom/auditrecords.go
+++ b/providers/namedotcom/auditrecords.go
@@ -14,11 +14,13 @@ import (
 func AuditRecords(records []*models.RecordConfig) []error {
 	a := rejectif.Auditor{}
 
+	a.Add("MX", rejectif.MxNull) // Last verified 2020-12-28
+
+	a.Add("SRV", rejectif.SrvHasNullTarget) // Last verified 2020-12-28
+
 	a.Add("TXT", MaxLengthNDC) // Last verified 2021-03-01
 
 	a.Add("TXT", rejectif.TxtIsEmpty) // Last verified 2021-03-01
-
-	a.Add("SRV", rejectif.SrvHasNullTarget) // Last verified 2020-12-28
 
 	return a.Audit(records)
 }