--- a/providers/netcup/auditrecords.go
+++ b/providers/netcup/auditrecords.go
@@ -11,6 +11,8 @@ import (
 func AuditRecords(records []*models.RecordConfig) []error {
 	a := rejectif.Auditor{}
 
+	a.Add("MX", rejectif.MxNull) // Last verified 2020-12-28
+
 	a.Add("TXT", rejectif.TxtIsEmpty) // Last verified 2021-03-01
 
 	return a.Audit(records)