--- a/providers/vultr/auditrecords.go
+++ b/providers/vultr/auditrecords.go
@@ -11,6 +11,8 @@ import (
 func AuditRecords(records []*models.RecordConfig) []error {
 	a := rejectif.Auditor{}
 
+	a.Add("MX", rejectif.MxNull) // Last verified 2020-12-28
+
 	a.Add("TXT", rejectif.TxtHasDoubleQuotes) // Last verified 2021-03-02
 	// Needs investigation. Could be a dnscontrol issue or
 	// the provider doesn't support double quotes.