This PR optimizes the integration tests.  Most are improved by 10-15% (about 30 seconds) but some (AZURE_DNS) have little improvement.


|NAME        |Original|Orig in Seconds|With Changes|NEW in Seconds|Improvement (seconds)|Improvement Percent|
|------------|--------|---------------|------------|--------------|---------------------|-------------------|
|GANDI_V5    |8:35    |515            |6:35        |395           |120                  |23%                |
|AZURE       |6:09    |369            |6:06        |366           |3                    |1%                 |
|CLOUDFLARE  |5:06    |306            |4:31        |271           |35                   |11%                |
|DIGITALOCEAN|4:07    |247            |3:49        |229           |18                   |7%                 |
|NAMEDOTCOM  |3:44    |224            |3:29        |209           |15                   |7%                 |
|GCLOUD      |3:34    |214            |3:05        |185           |29                   |14%                |
|POWERDNS    |3:19    |199            |2:54        |174           |25                   |13%                |
|ROUTE53     |2:58    |178            |2:31        |151           |27                   |15%                |
|HEXONET     |2:52    |172            |2:15        |135           |37                   |22%                |
|BIND        |1:08    |68             |1:07        |67            |1                    |1%                 |